small fix, cl-lib requirement
I wrote you on reddit about this, basically my emacs (27.2 then, now 28.05-gcc) froze when loading the theme. this should fix this. 
I could actually load the modeline, but not the theme itself or its functions with the :config call  (which is cool though that they can be used independently).
I'm sorry I don't describe the error more here, I just want to get this commit done for the moment to be able to use it... thought i had it done but it got lost  and just want it to work right now. basically i got a warning about the "for" variable being deprecated and could trace that to the cl-lib requirement missing, which is a pretty common error. there still were some messages about ivy, will report an issue if that happens again ;)